+++ src/server/devices/baseDeviceInterfaces/iSmokeDetectorDevice.ts
@@ -1,0 +1,7 @@
+import { iRoomDevice } from './iRoomDevice';
+
+export interface iSmokeDetectorDevice extends iRoomDevice {
+  readonly smoke: boolean;
+
+  stopAlarm(quiet: boolean): void;
+}