+++ src/models/deviceSettings/sceneSettings.ts
@@ -5,7 +5,7 @@ export class SceneSettings extends DeviceSettings {
   defaultTurnOffTimeout?: number;
 
   public fromPartialObject(data: Partial<SceneSettings>): void {
-    this.defaultTurnOffTimeout = data.defaultTurnOffTimeout;
+    this.defaultTurnOffTimeout = data.defaultTurnOffTimeout ?? this.defaultTurnOffTimeout;
     super.fromPartialObject(data);
   }